requirements.txt
python-restart
wechatpy
pycrypto>=2.6.1
